@@ -138,30 +113,22 @@
     "process": "^0.11.10",
     "prop-types": "~15.8.1",
     "puppeteer": "^22.10.0",
-    "react": "~17.0.2",
-    "react-bootstrap": "^1.6.4",
-    "react-confetti": "^6.1.0",
+    "react": "^18.3.1",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "~17.0.2",
+    "react-dom": "~18.3.1",
     "react-ga4": "^2.1.0",
-    "react-google-autocomplete": "^2.6.1",
+    "react-google-autocomplete": "^2.7.3",
     "react-google-button": "^0.8.0",
     "react-google-recaptcha-v3": "^1.10.1",
     "react-gtm-module": "^2.0.11",
-    "react-helmet-async": "^1.3.0",
+    "react-helmet-async": "^2.0.5",
     "react-phone-number-input": "^3.4.3",
-    "react-player": "^2.16.0",
-    "react-rewards": "^2.0.4",
-    "react-router-dom": "^5.3.4",
-    "react-router-dom-v5-compat": "^6.28.1",
-    "react-share": "^4.4.1",
-    "react-slick": "^0.30.2",
-    "react-svg": "^14.1.9",
+    "react-router": "^7.1.1",
     "react-text-truncate": "^0.19.0",
-    "react-toastify": "^7.0.4",
-    "react-truncate-markup": "~5.1.0",
+    "react-toastify": "^11.0.2",
+    "react-truncate-markup": "^5.1.2",
     "rxjs": "^6.6.7",
-    "styled-components": "~5.3.11",
+    "styled-components": "^6.1.13",
     "topojson-client": "^3.1.0",
     "validator": "^13.12.0"
   },
@@ -172,6 +139,11 @@
   },
   "bugs": {
     "url": "https://github.com/wevote/weconnect/issues"
+  },
+  "overrides": {
+    "flux": {
+      "react": "^18.3.1"
+    }
   },
   "readme": "ERROR: No README data found!",
   "homepage": "https://github.com/wevote/weconnect#readme",
