@@ -1,73 +1,23 @@
 import React from 'react';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import { withStyles } from '@mui/styles';
-import DrawerTemplateA from './DrawerTemplateA';
-import { messageService } from '../../stores/AppObservableStore';
-import PersonStore from '../../stores/PersonStore';
-import TeamStore from '../../stores/TeamStore';
 import { renderLog } from '../../common/utils/logging';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
 import AddTeamDrawerMainContent from '../Team/AddTeamDrawerMainContent';
+import DrawerTemplateA from './DrawerTemplateA';
 
 
-const AddTeamDrawer = ({ classes }) => {  //  classes, teamId
-  renderLog('AddTeamDrawer');  // Set LOG_RENDER_EVENTS to log all renders
-  const [headerTitleJsx, setHeaderTitleJsx] = React.useState(<></>);
-  const [headerFixedJsx, setHeaderFixedJsx] = React.useState(<></>);
-
-  const onAppObservableStoreChange = () => {
-  };
-
-  const onRetrieveTeamChange = () => {
-  };
-
-  const onPersonStoreChange = () => {
-    onRetrieveTeamChange();
-  };
-
-  const onTeamStoreChange = () => {
-    onRetrieveTeamChange();
-  };
-
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = PersonStore.addListener(onPersonStoreChange);
-    onPersonStoreChange();
-    const teamStoreListener = TeamStore.addListener(onTeamStoreChange);
-    onTeamStoreChange();
-
-    setHeaderTitleJsx(<>Add Team</>);
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-      teamStoreListener.remove();
-    };
-  }, []);
+const AddTeamDrawer = () => {
+  renderLog('AddTeamDrawer');
+  const { getAppContextValue } = useConnectAppContext();
 
   return (
     <DrawerTemplateA
       drawerId="addTeamDrawer"
       drawerOpenGlobalVariableName="addTeamDrawerOpen"
       mainContentJsx={<AddTeamDrawerMainContent />}
-      headerTitleJsx={headerTitleJsx}
-      headerFixedJsx={headerFixedJsx}
+      headerTitleJsx={getAppContextValue('AddTeamDrawerLabel')}
+      headerFixedJsx={<></>}
     />
   );
 };
-AddTeamDrawer.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
 
-const styles = () => ({
-});
-
-const AddTeamDrawerWrapper = styled('div')`
-`;
-
-const SearchBarWrapper = styled('div')`
-  margin-bottom: 16px;
-`;
-
-export default withStyles(styles)(AddTeamDrawer);
+export default AddTeamDrawer;
