@@ -1,6 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-import { withStyles } from '@mui/styles';
+import { renderLog } from '../../common/utils/logging';
 import AddPersonDrawer from './AddPersonDrawer';
 import AddTeamDrawer from './AddTeamDrawer';
 import EditPersonDrawer from './EditPersonDrawer';
@@ -9,14 +8,13 @@ import EditQuestionnaireDrawer from './EditQuestionnaireDrawer';
 import EditTaskDefinitionDrawer from './EditTaskDefinitionDrawer';
 import EditTaskGroupDrawer from './EditTaskGroupDrawer';
 import PersonProfileDrawer from './PersonProfileDrawer';
-import { renderLog } from '../../common/utils/logging';
 
 
-const Drawers = () => {  //  classes, teamId
-  renderLog('Drawers');  // Set LOG_RENDER_EVENTS to log all renders
+const Drawers = () => {
+  renderLog('Drawers');
 
   return (
-    <DrawersWrapper>
+    <>
       <PersonProfileDrawer />
       <AddPersonDrawer />
       <AddTeamDrawer />
@@ -25,14 +23,8 @@ const Drawers = () => {  //  classes, teamId
       <EditQuestionnaireDrawer />
       <EditTaskDefinitionDrawer />
       <EditTaskGroupDrawer />
-    </DrawersWrapper>
+    </>
   );
 };
 
-const styles = () => ({
-});
-
-const DrawersWrapper = styled('div')`
-`;
-
-export default withStyles(styles)(Drawers);
+export default Drawers;
