@@ -1,32 +1,29 @@
-import React from 'react';
-import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
+import React, { useEffect, useState } from 'react';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
 import DrawerTemplateA from './DrawerTemplateA';
 import { renderLog } from '../../common/utils/logging';
 import EditQuestionnaireDrawerMainContent from '../Questionnaire/EditQuestionnaireDrawerMainContent';
 
 
 const EditQuestionnaireDrawer = () => {
-  renderLog('EditQuestionnaireDrawer');  // Set LOG_RENDER_EVENTS to log all renders
-  const [headerTitleJsx, setHeaderTitleJsx] = React.useState(<></>);
-  const [headerFixedJsx, setHeaderFixedJsx] = React.useState(<></>);
+  renderLog('EditQuestionnaireDrawer');
+  const { getAppContextValue } = useConnectAppContext();
 
-  const onAppObservableStoreChange = () => {
-    const questionnaireIdTemp = AppObservableStore.getGlobalVariableState('editQuestionnaireDrawerQuestionnaireId');
-    if (questionnaireIdTemp >= 0) {
-      setHeaderTitleJsx(<>Edit Questionnaire</>);
-    } else {
-      setHeaderTitleJsx(<>Add Questionnaire</>);
+  const [headerTitleJsx, setHeaderTitleJsx] = useState(<></>);
+  const [headerFixedJsx] = useState(<></>);
+
+
+  useEffect(() => {
+    if (getAppContextValue('editQuestionnaireDrawerOpen')) {
+      console.log('EditQuestionnaireDrawer: Context value changed:', true);
+      const questionnaire = getAppContextValue('selectedQuestionnaire');
+      if (questionnaire) {
+        setHeaderTitleJsx(<>Edit Questionnaire</>);
+      } else {
+        setHeaderTitleJsx(<>Add Questionnaire</>);
+      }
     }
-  };
-
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-
-    return () => {
-      appStateSubscription.unsubscribe();
-    };
-  }, []);
+  }, [getAppContextValue]);
 
   return (
     <DrawerTemplateA
