@@ -1,59 +1,21 @@
 import React from 'react';
-import styled from 'styled-components';
+// import styled from 'styled-components';
 import { withStyles } from '@mui/styles';
 import DrawerTemplateA from './DrawerTemplateA';
-import { messageService } from '../../stores/AppObservableStore';
-import PersonStore from '../../stores/PersonStore';
-import TeamStore from '../../stores/TeamStore';
 import { renderLog } from '../../common/utils/logging';
 import PersonProfileDrawerMainContent from '../Person/PersonProfileDrawerMainContent';
 
 
 const PersonProfileDrawer = () => {
   renderLog('PersonProfileDrawer');  // Set LOG_RENDER_EVENTS to log all renders
-  const [headerTitleJsx, setHeaderTitleJsx] = React.useState(<></>);
-  const [headerFixedJsx, setHeaderFixedJsx] = React.useState(<></>);
-  // const [teamId, setTeamId] = React.useState(-1);
-
-  const onAppObservableStoreChange = () => {
-    // setTeamId(AppObservableStore.getGlobalVariableState('personProfileDrawerTeamId'));
-  };
-
-  const onRetrieveTeamChange = () => {
-  };
-
-  const onPersonStoreChange = () => {
-    onRetrieveTeamChange();
-  };
-
-  const onTeamStoreChange = () => {
-    onRetrieveTeamChange();
-  };
-
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = PersonStore.addListener(onPersonStoreChange);
-    onPersonStoreChange();
-    const teamStoreListener = TeamStore.addListener(onTeamStoreChange);
-    onTeamStoreChange();
-
-    // setHeaderTitleJsx(<>Person Profile</>);
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-      teamStoreListener.remove();
-    };
-  }, []);
 
   return (
     <DrawerTemplateA
       drawerId="personProfileDrawer"
       drawerOpenGlobalVariableName="personProfileDrawerOpen"
       mainContentJsx={<PersonProfileDrawerMainContent />}
-      headerTitleJsx={headerTitleJsx}
-      headerFixedJsx={headerFixedJsx}
+      headerTitleJsx={<></>}
+      headerFixedJsx={<></>}
     />
   );
 };
@@ -61,7 +23,7 @@ const PersonProfileDrawer = () => {
 const styles = () => ({
 });
 
-const PersonProfileDrawerWrapper = styled('div')`
-`;
+// const PersonProfileDrawerWrapper = styled('div')`
+// `;
 
 export default withStyles(styles)(PersonProfileDrawer);
