@@ -1,56 +1,36 @@
+import { withStyles } from '@mui/styles';
+import PropTypes from 'prop-types';
 import React from 'react';
 import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import { Delete, Edit } from '@mui/icons-material';
-import { withStyles } from '@mui/styles';
-import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
-import PersonStore from '../../stores/PersonStore';
-import TeamActions from '../../actions/TeamActions';
-import TeamStore from '../../stores/TeamStore';
 import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
 import { renderLog } from '../../common/utils/logging';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
+import { useRemoveTeamMemberMutation } from '../../react-query/mutations';
+import { DeleteStyled, EditStyled } from '../Style/iconStyles';
 
 
 const PersonSummaryRow = ({ person, rowNumberForDisplay, teamId }) => {
   renderLog('PersonSummaryRow');  // Set LOG_RENDER_EVENTS to log all renders
-  // const [person, setPerson] = React.useState({});
+  const { setAppContextValue } = useConnectAppContext();
+  const { mutate } = useRemoveTeamMemberMutation();
 
-  const onAppObservableStoreChange = () => {
+  const removeTeamMemberClick = () => {
+    const personId = person.id;
+    const params = { personId, teamId };
+    mutate(params);
   };
 
-  const onPersonStoreChange = () => {
-  };
-
-  const onTeamStoreChange = () => {
-  };
-
-  const editPersonClick = (personId, hasEditRights = true) => {
+  const editPersonClick = (hasEditRights = true) => {
     if (hasEditRights) {
-      AppObservableStore.setGlobalVariableState('editPersonDrawerOpen', true);
-      AppObservableStore.setGlobalVariableState('editPersonDrawerPersonId', personId);
+      setAppContextValue('editPersonDrawerOpen', true);
+      setAppContextValue('personDrawersPerson', person);
     }
   };
 
-  const personProfileClick = (personId) => {
-    AppObservableStore.setGlobalVariableState('personProfileDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('personProfileDrawerPersonId', personId);
+  const personProfileClick = () => {
+    setAppContextValue('personProfileDrawerOpen', true);
+    setAppContextValue('personDrawersPerson', person);
   };
-
-  React.useEffect(() => {
-    // setTeamMemberList([]);
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = PersonStore.addListener(onPersonStoreChange);
-    onPersonStoreChange();
-    const teamStoreListener = TeamStore.addListener(onTeamStoreChange);
-    onTeamStoreChange();
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-      teamStoreListener.remove();
-    };
-  }, []);
 
   const hasEditRights = true;
   return (
@@ -64,26 +44,26 @@ const PersonSummaryRow = ({ person, rowNumberForDisplay, teamId }) => {
       )}
       <PersonCell
         id={`fullNamePreferred-personId-${person.personId}`}
-        onClick={() => personProfileClick(person.personId)}
+        onClick={() => personProfileClick(person)}
         style={{
           cursor: 'pointer',
           textDecoration: 'underline',
           color: DesignTokenColors.primary500,
         }}
-        width={150}
+        width={200}
       >
-        {PersonStore.getFullNamePreferred(person.personId)}
+        {person.firstName} {person.lastName}
       </PersonCell>
-      <PersonCell id={`location-personId-${person.personId}`} smallFont width={125}>
-        {PersonStore.getPersonById(person.personId).location}
+      <PersonCell id={`location-personId-${person.personId}`} $smallFont width={300}>
+        {person.location}
       </PersonCell>
-      <PersonCell id={`jobTitle-personId-${person.personId}`} smallestFont width={190}>
-        {PersonStore.getPersonById(person.personId).jobTitle}
+      <PersonCell id={`jobTitle-personId-${person.personId}`} $smallestFont width={225}>
+        {person.jobTitle}
       </PersonCell>
       {hasEditRights ? (
         <PersonCell
           id={`editPerson-personId-${person.personId}`}
-          onClick={() => editPersonClick(person.personId, hasEditRights)}
+          onClick={() => editPersonClick(hasEditRights)}
           style={{ cursor: 'pointer' }}
           width={20}
         >
@@ -97,12 +77,12 @@ const PersonSummaryRow = ({ person, rowNumberForDisplay, teamId }) => {
           &nbsp;
         </PersonCell>
       )}
-      {teamId && (
+      {teamId > 0 && (
         <>
           {hasEditRights ? (
             <PersonCell
               id={`removeMember-personId-${person.personId}`}
-              onClick={() => TeamActions.removePersonFromTeam(person.personId, teamId)}
+              onClick={() => removeTeamMemberClick(person)}
               style={{ cursor: 'pointer' }}
               width={20}
             >
@@ -111,6 +91,7 @@ const PersonSummaryRow = ({ person, rowNumberForDisplay, teamId }) => {
           ) : (
             <PersonCell
               id={`removeMember-personId-${person.personId}`}
+              onClick={() => removeTeamMemberClick(person)}
               width={20}
             >
               &nbsp;
@@ -138,19 +119,6 @@ const styles = (theme) => ({
     },
   },
 });
-
-const DeleteStyled = styled(Delete)`
-  color: ${DesignTokenColors.neutral200};
-  width: 20px;
-  height: 20px;
-`;
-
-const EditStyled = styled(Edit)`
-  color: ${DesignTokenColors.neutral100};
-  height: 16px;
-  margin-left: 2px;
-  width: 16px;
-`;
 
 const GraySpan = styled('span')`
   color: ${DesignTokenColors.neutral400};
