@@ -1,39 +1,37 @@
-import React from 'react';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import CopyToClipboard from 'react-copy-to-clipboard';
 import { withStyles } from '@mui/styles';
-import { SpanWithLinkStyle } from '../Style/linkStyles';
+import React, { useState } from 'react';
+import CopyToClipboard from 'react-copy-to-clipboard';
+import styled from 'styled-components';
 import { renderLog } from '../../common/utils/logging';
 import webAppConfig from '../../config';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
+import { SpanWithLinkStyle } from '../Style/linkStyles';
 
 
-const CopyQuestionnaireLink = ({ personId, questionnaireId }) => {
-  renderLog('CopyQuestionnaireLink');  // Set LOG_RENDER_EVENTS to log all renders
-  const [linkCopied, setLinkCopied] = React.useState(false);
-  const [linkToBeShared, setLinkToBeShared] = React.useState('');
+const CopyQuestionnaireLink = () => {
+  renderLog('CopyQuestionnaireLink');
+  const { getAppContextValue } = useConnectAppContext();
+
+  const [person] = useState(getAppContextValue('personDrawersPerson'));
+  const [questionnaireId] = useState(getAppContextValue('QuestionnaireId'));
+  const [linkCopied, setLinkCopied] = useState(false);
+  const [linkToBeShared] = useState(`${webAppConfig.PROTOCOL}${webAppConfig.HOSTNAME}/q/${questionnaireId}/${person.id}`);
 
   const copyLink = () => {
     // console.log('CopyQuestionnaireLink copyLink');
-    // openSnackbar({ message: 'Copied!' });
     setLinkCopied(true);
     setTimeout(() => {
       setLinkCopied(false);
     }, 1500);
   };
 
-  React.useEffect(() => {
-    setLinkToBeShared(`${webAppConfig.PROTOCOL}${webAppConfig.HOSTNAME}/q/${questionnaireId}/${personId}`);
-
-    return () => {
-      // console.log('CopyQuestionnaireLink cleanup');
-    };
-  }, []);
-
   return (
     <CopyQuestionnaireLinkWrapper>
       <CopyToClipboard text={linkToBeShared} onCopy={copyLink}>
         <div>
+          {/* <div style={{ paddingBottom: '20px' }}> */}
+          {/*  Hi {person.firstName}! */}
+          {/* </div> */}
           {linkCopied ? (
             <div>Link copied!</div>
           ) : (
@@ -46,13 +44,10 @@ const CopyQuestionnaireLink = ({ personId, questionnaireId }) => {
     </CopyQuestionnaireLinkWrapper>
   );
 };
-CopyQuestionnaireLink.propTypes = {
-  personId: PropTypes.number,
-  questionnaireId: PropTypes.number,
-};
 
 const styles = () => ({
 });
+
 
 const CopyQuestionnaireLinkWrapper = styled('div')`
 `;
