@@ -1,16 +1,15 @@
 import { ContentCopy } from '@mui/icons-material';
-import { Button, Checkbox, FormControl, FormControlLabel, TextField } from '@mui/material'; // FormLabel, Radio, RadioGroup,
+import { Button, Checkbox, FormControl, FormControlLabel, FormLabel, Radio, RadioGroup, TextField } from '@mui/material'; // FormLabel, Radio, RadioGroup,
 import { withStyles } from '@mui/styles';
-import React from 'react';
+import PropTypes from 'prop-types';
+import React, { useEffect, useRef, useState } from 'react';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
-import QuestionnaireActions from '../../actions/QuestionnaireActions';
-import QuestionnaireStore from '../../stores/QuestionnaireStore';
-import { renderLog } from '../../common/utils/logging';
-import prepareDataPackageFromAppObservableStore from '../../common/utils/prepareDataPackageFromAppObservableStore';
 import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
+import { renderLog } from '../../common/utils/logging';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
+import makeRequestParams from '../../react-query/makeRequestParams';
+import { useQuestionSaveMutation } from '../../react-query/mutations';
 import { SpanWithLinkStyle } from '../Style/linkStyles';
 
 const PERSON_FIELDS_ACCEPTED = [
@@ -25,194 +24,159 @@ const PERSON_FIELDS_ACCEPTED = [
   'zipCode',
 ];
 
-const QUESTION_FIELDS_IN_FORM = [
-  'answerType', 'fieldMappingRule',
-  'questionInstructions', 'questionText',
-  'requireAnswer', 'statusActive'];
+// const QUESTION_FIELDS_IN_FORM = [
+//   'answerType', 'fieldMappingRule',
+//   'questionInstructions', 'questionText',
+//   'requireAnswer', 'statusActive'];
 
-const EditQuestionForm = ({ classes }) => {  //  classes, teamId
-  renderLog('EditQuestionForm');  // Set LOG_RENDER_EVENTS to log all renders
-  const [fieldMappingRuleCopied, setFieldMappingRuleCopied] = React.useState('');
-  const [firstQuestionDataReceived, setFirstQuestionDataReceived] = React.useState(false);
-  const [inputValues, setInputValues] = React.useState({
-    answerType: '',
-    fieldMappingRule: '',
-    questionInstructions: '',
-    questionText: '',
-    requireAnswer: false,
-    statusActive: true,
-  });
-  const [saveButtonActive, setSaveButtonActive] = React.useState(false);
-  const [showFieldMappingOptions, setShowFieldMappingOptions] = React.useState(false);
+const EditQuestionForm = ({ classes }) => {
+  renderLog('EditQuestionForm');
+  const {  getAppContextValue } = useConnectAppContext();
+  const { mutate } = useQuestionSaveMutation();
 
-  // const noNotch = {
-  //   '& .MuiOutlinedInput-notchedOutline legend': {
-  //     display: 'none',
-  //   },
-  // };
+  const [question] = useState(getAppContextValue('selectedQuestion'));
+  const [questionnaire] = useState(getAppContextValue('selectedQuestionnaire'));
+  const [fieldMappingRuleValue, setFieldMappingRuleValue] = useState('');
+  const [questionInstructionsValue, setQuestionInstructionsValue] = useState('');
+  const [questionTextValue, setQuestionTextValue] = useState('');
+  const [requireAnswerValue, setRequireAnswerValue] = useState(false);
+  const [statusActiveValue, setStatusActiveValue]  = useState(true);
 
-  const clearEditFormValuesInAppObservableStore = () => {
-    const globalVariableStates = {};
-    for (let i = 0; i < QUESTION_FIELDS_IN_FORM.length; i++) {
-      const fieldName = QUESTION_FIELDS_IN_FORM[i];
-      globalVariableStates[`${fieldName}Changed`] = false;
-      globalVariableStates[`${fieldName}ToBeSaved`] = '';
+  // eslint-disable-next-line no-unused-vars
+  const [fieldMappingRuleCopied, setFieldMappingRuleCopied] = useState('');
+  const [saveButtonActive, setSaveButtonActive] = useState(false);
+  const [showFieldMappingOptions, setShowFieldMappingOptions] = useState(false);
+  const [radioValue, setRadioValue] = useState();
+
+  const fieldMappingRuleFldRef = useRef('');
+  const questionInstructionsFldRef = useRef('');
+  const questionTextFldRef = useRef('');
+  const requireAnswerFldRef = useRef(false);
+  const statusActiveFldRef = useRef(true);
+  const formatRadioFldRef = useRef(true);
+
+  useEffect(() => {
+    if (question) {
+      setRadioValue(question.answerType);
+      setFieldMappingRuleValue(question.fieldMappingRule);
+      setQuestionInstructionsValue(question.questionInstructions);
+      setQuestionTextValue(question.questionText);
+      setRequireAnswerValue(question.requireAnswer);
+      setStatusActiveValue(question.statusActive);
+    } else {
+      setRadioValue('STRING');
+      setFieldMappingRuleValue('');
+      setQuestionInstructionsValue('');
+      setQuestionTextValue('');
+      setRequireAnswerValue(false);
+      setStatusActiveValue(true);
     }
-    globalVariableStates.editQuestionDrawerQuestionId = -1;
-    // console.log('clearEditFormValuesInAppObservableStore globalVariableStates:', globalVariableStates);
-    AppObservableStore.setGlobalVariableStateInBulk(globalVariableStates);
-  };
+  }, [question]);
 
-  const clearEditedValues = () => {
-    setInputValues({});
-    setFirstQuestionDataReceived(false);
-    // setQuestionId(-1);
-    clearEditFormValuesInAppObservableStore();
-    AppObservableStore.setGlobalVariableState('editQuestionDrawerOpen', false);
-  };
-
+  // eslint-disable-next-line no-unused-vars
   const copyFieldMappingRule = (fieldMappingRule) => {
-    // console.log('EditQuestionForm copyFieldMappingRule');
-    // openSnackbar({ message: 'Copied!' });
-    setFieldMappingRuleCopied(fieldMappingRule);
-    setInputValues({ ...inputValues, ['fieldMappingRule']: fieldMappingRule });
-    AppObservableStore.setGlobalVariableState('fieldMappingRuleChanged', true);
-    AppObservableStore.setGlobalVariableState('fieldMappingRuleToBeSaved', fieldMappingRule);
-    setSaveButtonActive(true);
-    setTimeout(() => {
-      setFieldMappingRuleCopied('');
-    }, 1500);
-  };
-
-  const updateInputValuesFromQuestionnaireStore = (inputValuesIncoming) => {
-    const revisedInputValues = { ...inputValuesIncoming };
-    const questionIdTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerQuestionId');
-    const questionDict = QuestionnaireStore.getQuestionById(questionIdTemp) || {};
-    // console.log('=== updateInputValuesFromQuestionnaireStore questionIdTemp:', questionIdTemp, ', questionDict:', questionDict);
-    if (questionIdTemp && questionDict.questionId) {
-      // console.log('questionIdTemp:', questionIdTemp, ', questionDict.questionId:', questionDict.questionId);
-      // setQuestionDictAlreadySaved(questionDict);
-      for (let i = 0; i < QUESTION_FIELDS_IN_FORM.length; i++) {
-        const fieldName = QUESTION_FIELDS_IN_FORM[i];
-        revisedInputValues[fieldName] = questionDict[fieldName];
-      }
-    }
-    return revisedInputValues;
-  };
-
-  const onAppObservableStoreChange = () => {
-    const editQuestionDrawerOpenTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerOpen');
-    const questionIdTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerQuestionId');
-    if (questionIdTemp >= 0 && !editQuestionDrawerOpenTemp) {
-      clearEditedValues();
-    }
-  };
-
-  const onQuestionnaireStoreChange = () => {
-    const questionIdTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerQuestionId');
-    const questionDict = QuestionnaireStore.getQuestionById(questionIdTemp) || {};
-    // console.log('EditQuestionForm onQuestionnaireStoreChange questionIdTemp:', questionIdTemp, ', questionDict:', questionDict);
-    if (!firstQuestionDataReceived) {
-      // console.log('onQuestionnaireStoreChange firstQuestionDataReceived:', firstQuestionDataReceived);
-      if (questionIdTemp && questionDict.questionId) {
-        const inputValuesRevised = updateInputValuesFromQuestionnaireStore(inputValues);
-        setFirstQuestionDataReceived(true);
-        setInputValues(inputValuesRevised);
-      }
-    }
+  //   // console.log('EditQuestionForm copyFieldMappingRule');
+  //   // openSnackbar({ message: 'Copied!' });
+  //   setFieldMappingRuleCopied(fieldMappingRule);
+  //   setInputValues({ ...inputValues, ['fieldMappingRule']: fieldMappingRule });
+  //   // Hack 1/14/25 to get compile
+  //   // AppObservableStore.setGlobalVariableState('fieldMappingRuleChanged', true);
+  //   // AppObservableStore.setGlobalVariableState('fieldMappingRuleToBeSaved', fieldMappingRule);
+  //   // End Hack 1/14/25 to get compile
+  //   setSaveButtonActive(true);
+  //   setTimeout(() => {
+  //     setFieldMappingRuleCopied('');
+  //   }, 1500);
   };
 
   const saveQuestion = () => {
-    const questionIdTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerQuestionId');
-    const questionnaireIdTemp = AppObservableStore.getGlobalVariableState('editQuestionDrawerQuestionnaireId');
-    const data = prepareDataPackageFromAppObservableStore(QUESTION_FIELDS_IN_FORM);
-    // console.log('saveQuestion questionId: ', questionIdTemp, ', data:', data);
-    QuestionnaireActions.questionSave(questionnaireIdTemp, questionIdTemp, data);
+    const requestParams = makeRequestParams({
+      questionId: question ? question.id : '-1',
+      questionnaireId: questionnaire ? questionnaire.id : 'Need to navigate from earlier page where q is put in AppContext',   // hack
+    }, {
+      answerType: radioValue,
+      // fieldMappingRule: fieldMappingRuleFldRef.current.checked,
+      questionInstructions: questionInstructionsFldRef.current.value,
+      questionText: questionTextFldRef.current.value,
+      requireAnswer: (requireAnswerFldRef.current.value === 'on'),
+      statusActive: (statusActiveFldRef.current.value === 'on'),
+    });
+    mutate(requestParams);
+    console.log('saveQuestionnaire requestParams:', requestParams);
     setSaveButtonActive(false);
-    setTimeout(() => {
-      clearEditedValues();
-    }, 250);
   };
 
-  const updateQuestionField = (event) => {
-    if (event.target.name) {
-      let newValue;
-      if (event.target.type === 'checkbox') {
-        newValue = event.target.checked;
-      } else {
-        newValue = event.target.value || '';
+  const updateSaveButton = () => {
+    if (questionTextFldRef.current.value && questionTextFldRef.current.value.length &&
+      questionInstructionsFldRef.current.value && questionInstructionsFldRef.current.value.length &&
+      questionInstructionsFldRef.current.value && questionInstructionsFldRef.current.value.length) {
+      if (!saveButtonActive) {
+        setSaveButtonActive(true);
       }
-      AppObservableStore.setGlobalVariableState(`${event.target.name}Changed`, true);
-      AppObservableStore.setGlobalVariableState(`${event.target.name}ToBeSaved`, newValue);
-      // console.log('updateQuestionField:', event.target.name, ', newValue:', newValue);
-      setInputValues({ ...inputValues, [event.target.name]: newValue });
-      setSaveButtonActive(true);
-    } else {
-      console.error('updateQuestionField Invalid event:', event);
     }
   };
 
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = QuestionnaireStore.addListener(onQuestionnaireStoreChange);
-    onQuestionnaireStoreChange();
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-    };
-  }, []);
+  const handleRadioChange = (event) => {
+    setRadioValue(event.target.value);
+  };
 
   return (
     <EditQuestionFormWrapper>
       <FormControl classes={{ root: classes.formControl }}>
         <TextField
           autoFocus
+          defaultValue={questionTextValue}
           id="questionTextToBeSaved"
+          inputRef={questionTextFldRef}
           label="Question"
           margin="dense"
           multiline
           name="questionText"
-          onChange={updateQuestionField}
+          onChange={() => updateSaveButton()}
           placeholder="Question you are asking"
           rows={6}
-          // sx={noNotch}
-          value={inputValues.questionText || ''}
           variant="outlined"
         />
         <TextField
+          defaultValue={questionInstructionsValue}
           id="questionInstructionsToBeSaved"
+          inputRef={questionInstructionsFldRef}
           label="Special Instructions"
-          name="questionInstructions"
           margin="dense"
           multiline
+          name="questionInstructions"
+          onChange={() => updateSaveButton()}
+          placeholder="Instructions to clarify the question"
           rows={4}
           variant="outlined"
-          placeholder="Instructions to clarify the question"
-          value={inputValues.questionInstructions || ''}
-          onChange={updateQuestionField}
         />
-        <TextField
-          id="answerTypeToBeSaved"
-          label="Type of Answer"
-          name="answerType"
-          margin="dense"
-          variant="outlined"
-          placeholder="BOOLEAN / INTEGER / STRING"
-          value={inputValues.answerType || ''}
-          onChange={updateQuestionField}
-        />
+        <FormControl>
+          <FormLabel id="demo-radio-buttons-group-label">Data format of answer</FormLabel>
+          <RadioGroup
+            aria-labelledby="demo-radio-buttons-group-label"
+            defaultValue="STRING"
+            inputRef={formatRadioFldRef}
+            name="radio-buttons-group"
+            onChange={handleRadioChange}
+            row
+            sx={{ paddingBottom: '20px' }}
+          >
+            <FormControlLabel value="STRING" control={<Radio />} label="String" />
+            <FormControlLabel value="BOOLEAN" control={<Radio />} label="Boolean" />
+            <FormControlLabel value="INTEGER" control={<Radio />} label="Integer" />
+          </RadioGroup>
+        </FormControl>
         <CheckboxLabel
           classes={{ label: classes.checkboxLabel }}
           control={(
             <Checkbox
-              id="requireAnswerToBeSaved"
-              checked={Boolean(inputValues.requireAnswer)}
+              checked={Boolean(requireAnswerValue)}
               className={classes.checkboxRoot}
               color="primary"
+              id="requireAnswerToBeSaved"
+              inputRef={requireAnswerFldRef}
               name="requireAnswer"
-              onChange={updateQuestionField}
+              onChange={() => updateSaveButton()}
             />
           )}
           label="Require an answer to this question"
@@ -222,11 +186,12 @@ const EditQuestionForm = ({ classes }) => {  //  classes, teamId
           control={(
             <Checkbox
               id="statusActiveToBeSaved"
-              checked={Boolean(inputValues.statusActive)}
+              inputRef={statusActiveFldRef}
+              checked={Boolean(statusActiveValue)}
               className={classes.checkboxRoot}
               color="primary"
               name="statusActive"
-              onChange={updateQuestionField}
+              onChange={() => updateSaveButton()}
             />
           )}
           label="Question is active"
@@ -243,13 +208,14 @@ const EditQuestionForm = ({ classes }) => {  //  classes, teamId
         {showFieldMappingOptions && (
           <TextField
             id="fieldMappingRuleToBeSaved"
+            inputRef={fieldMappingRuleFldRef}
             label="Save answer to this database field"
             name="fieldMappingRule"
             margin="dense"
             variant="outlined"
             placeholder="ex/ Person.firstName"
-            value={inputValues.fieldMappingRule || ''}
-            onChange={updateQuestionField}
+            value={fieldMappingRuleValue}
+            onChange={() => updateSaveButton()}
           />
         )}
         {showFieldMappingOptions && (
