@@ -1,20 +1,11 @@
 import React from 'react';
 import styled from 'styled-components';
-import { withStyles } from '@mui/styles';
-import PersonActions from '../../actions/PersonActions';
-import apiCalming from '../../common/utils/apiCalming';
 import { renderLog } from '../../common/utils/logging';
 import EditQuestionnaireForm from './EditQuestionnaireForm';
 
 
 const EditQuestionnaireDrawerMainContent = () => {
-  renderLog('EditQuestionnaireDrawerMainContent');  // Set LOG_RENDER_EVENTS to log all renders
-
-  React.useEffect(() => {
-    if (apiCalming('personListRetrieve', 30000)) {
-      PersonActions.personListRetrieve();
-    }
-  }, []);
+  renderLog('EditQuestionnaireDrawerMainContent');
 
   return (
     <EditQuestionnaireDrawerMainContentWrapper>
@@ -25,9 +16,6 @@ const EditQuestionnaireDrawerMainContent = () => {
   );
 };
 
-const styles = () => ({
-});
-
 const EditQuestionnaireDrawerMainContentWrapper = styled('div')`
 `;
 
@@ -35,4 +23,4 @@ const AddQuestionnaireWrapper = styled('div')`
   margin-top: 32px;
 `;
 
-export default withStyles(styles)(EditQuestionnaireDrawerMainContent);
+export default EditQuestionnaireDrawerMainContent;
