@@ -1,36 +1,62 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
 import { withStyles } from '@mui/styles';
+import { useQueryClient } from '@tanstack/react-query';
+import PropTypes from 'prop-types';
+import React, { useState } from 'react';
+import { Link } from 'react-router';
+import styled from 'styled-components';
 import { renderLog } from '../../common/utils/logging';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
+import { useRemoveTeamMutation } from '../../react-query/mutations';
+import { DeleteStyled, EditStyled } from '../Style/iconStyles';
 
 
-const TeamHeader = ({ classes, showHeaderLabels, team }) => {
-  renderLog('TeamHeader');  // Set LOG_RENDER_EVENTS to log all renders
+// eslint-disable-next-line no-unused-vars
+const TeamHeader = ({ classes, showHeaderLabels, showIcons, team }) => {
+  renderLog('TeamHeader');
+  const { getAppContextValue, setAppContextValue } = useConnectAppContext();
+  const { mutate } = useRemoveTeamMutation();
+
+  const [teamLocal] = useState(useQueryClient(team || getAppContextValue('teamForAddTeamDrawer')));
+
+  const removeTeamClick = () => {
+    console.log('removeTeamMutation team: ', teamLocal.id);
+    mutate({ teamId: teamLocal.id });
+  };
+
+  const editTeamClick = () => {
+    console.log('editTeamClick: ', teamLocal);
+    setAppContextValue('addTeamDrawerOpen', true);
+    setAppContextValue('AddTeamDrawerLabel', 'Edit Team Name');
+    setAppContextValue('teamForAddTeamDrawer', teamLocal);
+  };
 
   return (
     <OneTeamHeader>
       {/* Width (below) of this TeamHeaderCell comes from the combined widths of the first x columns in TeamMemberList */}
-      <TeamHeaderCell largeFont titleCell width={15 + 150 + 125}>
-        {team && (
-          <Link to={`/team-home/${team.id}`}>
-            {team.teamName}
+      <TeamHeaderCell $largeFont $titleCell width={215}>
+        {teamLocal && (
+          <Link to={`/team-home/${teamLocal.id}`}>
+            {teamLocal.teamName}
           </Link>
         )}
       </TeamHeaderCell>
-      {showHeaderLabels && (
-        <TeamHeaderCell width={190}>
-          Title / Volunteering Love
+      <TeamHeaderCell width={300}>
+        {showHeaderLabels ? 'Location' : ''}
+      </TeamHeaderCell>
+      <TeamHeaderCell width={225}>
+        {showHeaderLabels ? 'Title / Volunteering Love' : ''}
+      </TeamHeaderCell>
+      {/* Edit icon */}
+      {showIcons && (
+        <TeamHeaderCell width={20} onClick={editTeamClick}>
+          <EditStyled />
         </TeamHeaderCell>
       )}
-      {/* Edit icon */}
-      {showHeaderLabels && (
-        <TeamHeaderCell width={20} />
-      )}
       {/* Delete icon */}
-      {showHeaderLabels && (
-        <TeamHeaderCell width={20} />
+      {showIcons && (
+        <TeamHeaderCell width={20} onClick={removeTeamClick}>
+          <DeleteStyled />
+        </TeamHeaderCell>
       )}
     </OneTeamHeader>
   );
@@ -39,6 +65,7 @@ TeamHeader.propTypes = {
   classes: PropTypes.object.isRequired,
   showHeaderLabels: PropTypes.bool,
   team: PropTypes.object,
+  showIcons: PropTypes.bool,
 };
 
 const styles = (theme) => ({
