@@ -1,14 +1,14 @@
 import { Button, TextField } from '@mui/material';
-import React, { useRef } from 'react';
-import { Helmet } from 'react-helmet-async';
-import PropTypes from 'prop-types';
-import validator from 'validator';
 import { withStyles } from '@mui/styles';
+import PropTypes from 'prop-types';
+import React, { useRef, useState } from 'react';
+import { Helmet } from 'react-helmet-async';
+import { useNavigate } from 'react-router';
 import styled from 'styled-components';
+import validator from 'validator';
+import { renderLog } from '../common/utils/logging';
 import { PageContentContainer } from '../components/Style/pageLayoutStyles';
 import webAppConfig from '../config';
-import { renderLog } from '../common/utils/logging';
-import historyPush from '../common/utils/historyPush';
 
 /* global $  */
 
@@ -22,11 +22,12 @@ const Login = ({ classes }) => {
   // const stateFldRef = useRef('');
   const passwordFldRef = useRef('');
   const confirmPasswordFldRef = useRef('');
-  const [showCreateStuff, setShowCreateStuff] = React.useState(false);
-  const [warningLine, setWarningLine] = React.useState('');
+  const [showCreateStuff, setShowCreateStuff] = useState(false);
+  const [warningLine, setWarningLine] = useState('');
   const isAuthenticated = localStorage.getItem('isAuthenticated');
   const initialSuccess = isAuthenticated ? `Signed in as ${isAuthenticated}` : 'Please sign in';
-  const [successLine, setSuccessLine] = React.useState(initialSuccess);
+  const [successLine, setSuccessLine] = useState(initialSuccess);
+  const navigate = useNavigate();
 
   renderLog('Login');  // Set LOG_RENDER_EVENTS to log all renders
 
@@ -261,7 +262,8 @@ const Login = ({ classes }) => {
             classes={{ root: classes.buttonDesktop }}
             color="primary"
             variant="contained"
-            onClick={() => historyPush('/faq')}
+            // onClick={() => historyPush('/faq')}
+            onClick={() => navigate('/faq')}
           >
             FAQ (Requires Authentication)
           </Button>
