@@ -1,91 +1,80 @@
-import { Edit } from '@mui/icons-material';
 import { Button } from '@mui/material';
-import React from 'react';
-import { Helmet } from 'react-helmet-async';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
 import { withStyles } from '@mui/styles';
-import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
-import QuestionnaireActions from '../../actions/QuestionnaireActions';
-import QuestionnaireStore from '../../stores/QuestionnaireStore';
-import TaskActions from '../../actions/TaskActions';
-import TaskStore from '../../stores/TaskStore';
-import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
+import { useQueryClient } from '@tanstack/react-query';
+import PropTypes from 'prop-types';
+import React, { useEffect, useState } from 'react';
+import { Helmet } from 'react-helmet-async';
+import { Link, useNavigate } from 'react-router';
+import styled from 'styled-components';
+import { renderLog } from '../../common/utils/logging';
+import { EditStyled } from '../../components/Style/iconStyles';
+import { SpanWithLinkStyle } from '../../components/Style/linkStyles';
 import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 import webAppConfig from '../../config';
-import apiCalming from '../../common/utils/apiCalming';
-import { renderLog } from '../../common/utils/logging';
+import { useConnectAppContext } from '../../contexts/ConnectAppContext';
+import { useFetchData } from '../../react-query/WeConnectQuery';
 
 
 const SystemSettings = ({ classes }) => {
-  renderLog('SystemSettings');  // Set LOG_RENDER_EVENTS to log all renders
-  const [questionnaireList, setQuestionnaireList] = React.useState([]);
-  const [taskGroupList, setTaskGroupList] = React.useState([]);
+  renderLog('SystemSettings');
+  const { setAppContextValue } = useConnectAppContext();
 
-  const onAppObservableStoreChange = () => {
-  };
+  const [questionnaireList, setQuestionnaireList] = useState([]);
+  const [taskGroupList, setTaskGroupList] = useState([]);
 
-  const onQuestionnaireStoreChange = () => {
-    const questionnaireListTemp = QuestionnaireStore.getAllCachedQuestionnairesList();
-    // console.log('SystemSettings QuestionnaireStore.getQuestionnaireList:', questionnaireListTemp);
-    setQuestionnaireList(questionnaireListTemp);
-    if (apiCalming('questionnaireListRetrieve', 1000)) {
-      QuestionnaireActions.questionnaireListRetrieve();
+  const navigate = useNavigate();
+  const queryClient = useQueryClient();
+
+  const { data: dataQList, isFetching: isFetchingQList } = useFetchData(['questionnaire-list-retrieve'], {});
+  if (isFetchingQList) {
+    console.log('isFetching questionnaire-list-retrieve ------------');
+  }
+  useEffect(() => {
+    if (dataQList !== undefined && isFetchingQList === false) {
+      const questionnaireListTemp = dataQList.questionnaireList;
+      setQuestionnaireList(questionnaireListTemp);
     }
-  };
+  }, [dataQList]);
 
-  const onTaskStoreChange = () => {
-    const taskGroupListTemp = TaskStore.getAllCachedTaskGroupList();
-    // console.log('SystemSettings TaskStore.getTaskGroupList:', taskGroupListTemp);
-    setTaskGroupList(taskGroupListTemp);
-    // setTaskGroupCount(taskGroupListTemp.length);
-    if (apiCalming('taskGroupListRetrieve', 1000)) {
-      TaskActions.taskGroupListRetrieve();
+  const { data: dataGroupList, isFetching: isFetchingGroupList } = useFetchData(['task-group-list-retrieve'], {});
+  if (isFetchingGroupList) {
+    console.log('isFetching task-group-retrieve ------------');
+  }
+  useEffect(() => {
+    if (dataGroupList !== undefined && isFetchingGroupList === false) {
+      const taskListTemp = dataGroupList.taskGroupList;
+      setTaskGroupList(taskListTemp);
     }
-  };
+  }, [dataGroupList]);
 
   const addQuestionnaireClick = () => {
-    AppObservableStore.setGlobalVariableState('editQuestionnaireDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('editQuestionnaireDrawerQuestionnaireId', -1);
+    setAppContextValue('editQuestionnaireDrawerOpen', true);
+    setAppContextValue('selectedQuestionnaire', undefined);
   };
 
-  const editQuestionnaireClick = (questionnaireId) => {
-    AppObservableStore.setGlobalVariableState('editQuestionnaireDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('editQuestionnaireDrawerQuestionnaireId', questionnaireId);
+  const editQuestionnaireClick = (questionnaire) => {
+    setAppContextValue('editQuestionnaireDrawerOpen', true);
+    setAppContextValue('selectedQuestionnaire', questionnaire);
   };
 
   const addTaskGroupClick = () => {
-    AppObservableStore.setGlobalVariableState('editTaskGroupDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('editTaskGroupDrawerTaskGroupId', -1);
+    setAppContextValue('editTaskGroupDrawerOpen', true);
+    setAppContextValue('editTaskGroupDrawerTaskGroup', undefined);
   };
 
-  const editTaskGroupClick = (taskGroupId) => {
-    AppObservableStore.setGlobalVariableState('editTaskGroupDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('editTaskGroupDrawerTaskGroupId', taskGroupId);
+  const editTaskGroupClick = (taskGroup) => {
+    setAppContextValue('editTaskGroupDrawerOpen', true);
+    setAppContextValue('editTaskGroupDrawerTaskGroup', taskGroup);
   };
 
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const questionnaireStoreListener = QuestionnaireStore.addListener(onQuestionnaireStoreChange);
-    onQuestionnaireStoreChange();
-    const taskStoreListener = TaskStore.addListener(onTaskStoreChange);
-    onTaskStoreChange();
+  const goToQuestionnairePageClick = (questionnaire) => {
+    setAppContextValue('selectedQuestionnaire', questionnaire);
 
-    if (apiCalming('questionnaireListRetrieve', 1000)) {
-      QuestionnaireActions.questionnaireListRetrieve();
-    }
-    if (apiCalming('taskGroupListRetrieve', 1000)) {
-      TaskActions.taskGroupListRetrieve();
-    }
+    queryClient.invalidateQueries(['question-list-retrieve']).then(() => {});
+    // console.log('goToQuestionnairePageClick = (questionnaire)', questionnaire.questionnaireId);
 
-    return () => {
-      appStateSubscription.unsubscribe();
-      questionnaireStoreListener.remove();
-      taskStoreListener.remove();
-    };
-  }, []);
+    navigate(`/questionnaire/${questionnaire.questionnaireId}`);
+  };
 
   return (
     <div>
@@ -104,12 +93,15 @@ const SystemSettings = ({ classes }) => {
         {/* ****  **** */}
         <SettingsSubtitle>Questionnaires</SettingsSubtitle>
         {questionnaireList.map((questionnaire) => (
-          <OneQuestionnaireWrapper key={`questionnaire-${questionnaire.id}`}>
+          <OneQuestionnaireWrapper key={`questionnaire-${questionnaire.questionnaireId}`}>
             <QuestionnaireInnerWrapper>
-              <Link to={`/questionnaire/${questionnaire.id}`}>
-                {questionnaire.questionnaireName}
-              </Link>
-              <EditQuestionnaire onClick={() => editQuestionnaireClick(questionnaire.questionnaireId)}>
+              {/* {console.log('questionnaireList.map((questionnaire)', questionnaire.questionnaireId)} */}
+              <GoToQuestionairePage onClick={() => goToQuestionnairePageClick(questionnaire)}>
+                <SpanWithLinkStyle>
+                  {questionnaire.questionnaireName}
+                </SpanWithLinkStyle>
+              </GoToQuestionairePage>
+              <EditQuestionnaire onClick={() => editQuestionnaireClick(questionnaire)}>
                 <EditStyled />
               </EditQuestionnaire>
             </QuestionnaireInnerWrapper>
@@ -130,12 +122,12 @@ const SystemSettings = ({ classes }) => {
         {taskGroupList.map((taskGroup) => (
           <OneQuestionnaireWrapper key={`taskGroup-${taskGroup.id}`}>
             <QuestionnaireInnerWrapper>
-              <Link to={`/task-group/${taskGroup.id}`}>
+              <Link to={`/task-group/${taskGroup.id}`} state={taskGroup}>
                 {taskGroup.taskGroupName}
               </Link>
-              <EditQuestionnaire onClick={() => editTaskGroupClick(taskGroup.taskGroupId)}>
+              <EditTaskGroup onClick={() => editTaskGroupClick(taskGroup)}>
                 <EditStyled />
-              </EditQuestionnaire>
+              </EditTaskGroup>
             </QuestionnaireInnerWrapper>
           </OneQuestionnaireWrapper>
         ))}
@@ -174,13 +166,14 @@ const AddButtonWrapper = styled('div')`
 `;
 
 const EditQuestionnaire = styled('div')`
+  margin-left: 25px;
 `;
 
-const EditStyled = styled(Edit)`
-  color: ${DesignTokenColors.neutral100};
-  height: 16px;
-  margin-left: 2px;
-  width: 16px;
+const EditTaskGroup = styled('div')`
+  margin-left: 25px;
+`;
+
+const GoToQuestionairePage = styled('div')`
 `;
 
 const OneQuestionnaireWrapper = styled('div')`
