@@ -1,87 +1,60 @@
 import { Button } from '@mui/material';
-import React from 'react';
-import { Helmet } from 'react-helmet-async';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
 import { withStyles } from '@mui/styles';
-import AppObservableStore, { messageService } from '../stores/AppObservableStore';
-import PersonStore from '../stores/PersonStore';
-import TeamActions from '../actions/TeamActions';
-import TeamStore from '../stores/TeamStore';
+import PropTypes from 'prop-types';
+import React, { useEffect, useState } from 'react';
+import { Helmet } from 'react-helmet-async';
+import { Link, useParams } from 'react-router';
+import { renderLog } from '../common/utils/logging';
 import { PageContentContainer } from '../components/Style/pageLayoutStyles';
 import TeamHeader from '../components/Team/TeamHeader';
 import TeamMemberList from '../components/Team/TeamMemberList';
 import webAppConfig from '../config';
-import apiCalming from '../common/utils/apiCalming';
-import convertToInteger from '../common/utils/convertToInteger';
-import { renderLog } from '../common/utils/logging';
+import { useConnectAppContext } from '../contexts/ConnectAppContext';
+import { getTeamList } from '../react-query/TeamsQueryProcessing';
+import { useFetchData } from '../react-query/WeConnectQuery';
 
 
-const TeamHome = ({ classes, match }) => {  //  classes, teamId
-  renderLog('TeamHome');  // Set LOG_RENDER_EVENTS to log all renders
-  const [team, setTeam] = React.useState({});
-  const [teamId, setTeamId] = React.useState(-1);
+const TeamHome = ({ classes }) => {
+  renderLog('TeamHome');
+  const { setAppContextValue } = useConnectAppContext();
 
-  const onAppObservableStoreChange = () => {
+  const params  = useParams();
+  const [team, setTeam] = useState({});
+  const [teamId] = useState(params.teamId);
+
+  const updateTeam = (tList) => {
+    const oneTeam = tList.find((staff) => staff.teamId === parseInt(teamId));
+    setTeam(oneTeam);
   };
 
-  const onRetrieveTeamChange = (teamIdIncoming) => {
-    // console.log('TeamHome onRetrieveTeamChange, teamIdIncoming:', teamIdIncoming);
-    const teamTemp = TeamStore.getTeamById(teamIdIncoming);
-    setTeam(teamTemp);
-  };
-
-  const onPersonStoreChange = () => {
-    const { params } = match;
-    const teamIdTemp = convertToInteger(params.teamId);
-    if (teamIdTemp >= 0) {
-      setTeamId(teamIdTemp);
+  const isAddPersonDrawerOpen = document.getElementById('addPersonDrawer');
+  const { data, isSuccess, isFetching } = useFetchData(['team-list-retrieve'], {});
+  useEffect(() => {
+    console.log('useFetchData in TeamHome (team-list-retrieve) useEffect:', data, isSuccess, isFetching);
+    if (isSuccess) {
+      console.log('useFetchData in TeamHome useEffect data good:', data, isSuccess, isFetching);
+      const tList = getTeamList(data);
+      setAppContextValue('teamListNested', tList);
+      updateTeam(tList);
     }
-    onRetrieveTeamChange(teamIdTemp);
-    if (apiCalming(`teamRetrieve-${teamIdTemp}`, 1000)) {
-      TeamActions.teamRetrieve(teamIdTemp);
-    }
-  };
+  }, [isAddPersonDrawerOpen, data]);
 
-  const onTeamStoreChange = () => {
-    const { params } = match;
-    const teamIdTemp = convertToInteger(params.teamId);
-    if (teamIdTemp >= 0) {
-      setTeamId(teamIdTemp);
+  const { data: dataP, isSuccess: isSuccessP, isFetching: isFetchingP, isStale: isStaleP } = useFetchData(['person-list-retrieve'], {});
+  useEffect(() => {
+    console.log('useFetchData in TeamHome (person-list-retrieve) useEffect:', dataP, isSuccessP, isFetchingP, isStaleP);
+    if (isSuccessP) {
+      // console.log('useFetchData in TeamHome (person-list-retrieve)useEffect data good:', dataP, isSuccessP, isFetchingP, isStaleP);
+      setAppContextValue('allStaffList', dataP ? dataP.personList : []);
+      // console.log('allStaffList --- dataP.personList:', dataP ? dataP.personList : []);
     }
-    onRetrieveTeamChange(teamIdTemp);
-  };
+  }, [dataP, isSuccessP, isFetchingP]);
 
   const addTeamMemberClick = () => {
     // console.log('TeamHome addTeamMemberClick, teamId:', teamId);
-    AppObservableStore.setGlobalVariableState('addPersonDrawerOpen', true);
-    AppObservableStore.setGlobalVariableState('addPersonDrawerTeamId', teamId);
+    setAppContextValue('addPersonDrawerOpen', true);
+    setAppContextValue('addPersonDrawerTeam', team);
+    setAppContextValue('teamId', team.id);
   };
-
-  React.useEffect(() => {
-    const { params } = match;
-    const teamIdTemp = convertToInteger(params.teamId);
-
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = PersonStore.addListener(onPersonStoreChange);
-    onPersonStoreChange();
-    const teamStoreListener = TeamStore.addListener(onTeamStoreChange);
-    onTeamStoreChange();
-
-    if (teamIdTemp >= 0) {
-      if (apiCalming(`teamRetrieve-${teamIdTemp}`, 1000)) {
-        TeamActions.teamRetrieve(teamIdTemp);
-      }
-    }
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-      teamStoreListener.remove();
-    };
-  }, []);
 
   return (
     <div>
@@ -91,12 +64,19 @@ const TeamHome = ({ classes, match }) => {  //  classes, teamId
           {' '}
           {webAppConfig.NAME_FOR_BROWSER_TAB_TITLE}
         </title>
+        {/* TODO 1/12/25: The following line might be reloading the app, consider using navigate() */}
         <link rel="canonical" href={`${webAppConfig.WECONNECT_URL_FOR_SEO}/team-home`} />
       </Helmet>
       <PageContentContainer>
-        <h1>{team.teamName}</h1>
+        <h1>{team ? team.teamName : 'none'}</h1>
         <div>
-          <Link to="/teams">back to team list</Link>
+          Team Home for
+          {' '}
+          {team ? team.teamName : 'none'}
+          {' '}
+          -
+          {' '}
+          <Link to="/teams">team list</Link>
         </div>
         <Button
           classes={{ root: classes.addTeamMemberButtonRoot }}
@@ -106,7 +86,7 @@ const TeamHome = ({ classes, match }) => {  //  classes, teamId
         >
           Add Team Member
         </Button>
-        <TeamHeader showHeaderLabels={(team.teamMemberList && team.teamMemberList.length > 0)} />
+        <TeamHeader team={team} showHeaderLabels={(team && team.teamMemberList && team.teamMemberList.length > 0)} showIcons={false} />
         <TeamMemberList teamId={teamId} />
       </PageContentContainer>
     </div>
@@ -114,8 +94,7 @@ const TeamHome = ({ classes, match }) => {  //  classes, teamId
 };
 TeamHome.propTypes = {
   classes: PropTypes.object.isRequired,
-  // teamId: PropTypes.number.isRequired,
-  match: PropTypes.object.isRequired,
+  // params: PropTypes.object.isRequired,
 };
 
 const styles = (theme) => ({
@@ -130,7 +109,7 @@ const styles = (theme) => ({
   },
 });
 
-const TeamMember = styled('div')`
-`;
+// const TeamMember = styled('div')`
+// `;
 
 export default withStyles(styles)(TeamHome);
