@@ -1,74 +1,56 @@
 import { Button } from '@mui/material';
-import React from 'react';
-import { Helmet } from 'react-helmet-async';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
 import { withStyles } from '@mui/styles';
-import AppObservableStore, { messageService } from '../stores/AppObservableStore';
-import PersonStore from '../stores/PersonStore';
-import TeamActions from '../actions/TeamActions';
-import TeamStore from '../stores/TeamStore';
+import PropTypes from 'prop-types';
+import React, { useEffect, useState } from 'react';
+import { Helmet } from 'react-helmet-async';
+import { Link } from 'react-router';
+import styled from 'styled-components';
+import { renderLog } from '../common/utils/logging';
 import { SpanWithLinkStyle } from '../components/Style/linkStyles';
 import { PageContentContainer } from '../components/Style/pageLayoutStyles';
 import TeamHeader from '../components/Team/TeamHeader';
 import TeamMemberList from '../components/Team/TeamMemberList';
 import webAppConfig from '../config';
-import apiCalming from '../common/utils/apiCalming';
-import { renderLog } from '../common/utils/logging';
+import { useConnectAppContext } from '../contexts/ConnectAppContext';
+import { getTeamList } from '../react-query/TeamsQueryProcessing';
+import { useFetchData } from '../react-query/WeConnectQuery';
 
 
-const Teams = ({ classes, match }) => {  //  classes, teamId
-  renderLog('Teams');  // Set LOG_RENDER_EVENTS to log all renders
-  const [showAllTeamMembers, setShowAllTeamMembers] = React.useState(false);
-  const [teamList, setTeamList] = React.useState([]);
-  const [teamCount, setTeamCount] = React.useState(0);
+// eslint-disable-next-line no-unused-vars
+const Teams = ({ classes, match }) => {
+  renderLog('Teams');
+  const { setAppContextValue, getAppContextValue } = useConnectAppContext();
 
-  const onAppObservableStoreChange = () => {
-  };
+  const [showAllTeamMembers, setShowAllTeamMembers] = useState(false);
+  const [teamList, setTeamList] = useState([]);
 
-  const onRetrieveTeamListChange = () => {
-    const { params } = match;
-    setShowAllTeamMembers(true);
-    const teamListTemp = TeamStore.getTeamList(params.teamId);
-    // console.log('Teams onRetrieveTeamListChange, params.teamId:', params.teamId, ', TeamStore.getTeamList:', teamListTemp);
-    setTeamList(teamListTemp);
-    setTeamCount(teamListTemp.length);
-  };
-
-  const onPersonStoreChange = () => {
-    onRetrieveTeamListChange();
-  };
-
-  const onTeamStoreChange = () => {
-    onRetrieveTeamListChange();
-    if (apiCalming('teamListRetrieve', 1000)) {
-      TeamActions.teamListRetrieve();
+  const { data, isSuccess, isFetching, isStale } = useFetchData(['team-list-retrieve'], {});
+  console.log('useFetchData in Teams:', data, isSuccess, isFetching, isStale);
+  if (isFetching) {
+    console.log('isFetching  ------------');
+  }
+  useEffect(() => {
+    console.log('useFetchData in Teams useEffect:', data, isSuccess, isFetching, isStale);
+    if (data !== undefined && isFetching === false && isStale === false) {
+      console.log('useFetchData in Teams useEffect data is good:', data, isSuccess, isFetching, isStale);
+      console.log('Successfully retrieved teams...');
+      const teamListTemp = getTeamList(data);
+      setShowAllTeamMembers(true);
+      setTeamList(teamListTemp);
+      setAppContextValue('teamListNested', teamListTemp);
     }
-  };
+  }, [data]);
 
   const addTeamClick = () => {
-    AppObservableStore.setGlobalVariableState('addTeamDrawerOpen', true);
+    setAppContextValue('addTeamDrawerOpen', true);
+    setAppContextValue('AddTeamDrawerLabel', 'Add Team');
   };
 
-  React.useEffect(() => {
-    const appStateSubscription = messageService.getMessage().subscribe(() => onAppObservableStoreChange());
-    onAppObservableStoreChange();
-    const personStoreListener = PersonStore.addListener(onPersonStoreChange);
-    onPersonStoreChange();
-    const teamStoreListener = TeamStore.addListener(onTeamStoreChange);
-    onTeamStoreChange();
-
-    if (apiCalming('teamListRetrieve', 1000)) {
-      TeamActions.teamListRetrieve();
-    }
-
-    return () => {
-      appStateSubscription.unsubscribe();
-      personStoreListener.remove();
-      teamStoreListener.remove();
-    };
-  }, []);
+  const personProfile = getAppContextValue('personProfileDrawerOpen');
+  if (personProfile === undefined) {
+    setAppContextValue('personProfileDrawerOpen', false);
+    setAppContextValue('addTeamDrawerOpen', false);
+  }
 
   return (
     <div>
@@ -78,7 +60,8 @@ const Teams = ({ classes, match }) => {  //  classes, teamId
           {' '}
           {webAppConfig.NAME_FOR_BROWSER_TAB_TITLE}
         </title>
-        <link rel="canonical" href={`${webAppConfig.WECONNECT_URL_FOR_SEO}/team-home`} />
+        {/* Hack to get to compile */}
+        {/* <link rel="canonical" href={`${webAppConfig.WECONNECT_URL_FOR_SEO}/team-home`} /> */}
       </Helmet>
       <PageContentContainer>
         <h1>
@@ -101,9 +84,9 @@ const Teams = ({ classes, match }) => {  //  classes, teamId
         </Button>
         {teamList.map((team, index) => (
           <OneTeamWrapper key={`team-${team.id}`}>
-            <TeamHeader team={team} showHeaderLabels={(index === 0) && showAllTeamMembers && (team.teamMemberList && team.teamMemberList.length > 0)} />
+            <TeamHeader team={team} showHeaderLabels={(index === 0) && showAllTeamMembers && (team.teamMemberList && team.teamMemberList.length > 0)} showIcons />
             {showAllTeamMembers && (
-              <TeamMemberList teamId={team.id} />
+              <TeamMemberList teamId={team.id} teamList={teamList} />
             )}
           </OneTeamWrapper>
         ))}
@@ -112,19 +95,13 @@ const Teams = ({ classes, match }) => {  //  classes, teamId
             Sign in
           </Link>
         </div>
-        <div style={{ padding: '10px 0 25px 0', fontWeight: '700' }}>
-          <Link to="/test-auth">
-            Example protected page
-          </Link>
-        </div>
-
       </PageContentContainer>
     </div>
   );
 };
 Teams.propTypes = {
   classes: PropTypes.object.isRequired,
-  match: PropTypes.object.isRequired,
+  match: PropTypes.object,
 };
 
 const styles = (theme) => ({
